Move Shuttle tests to the individual modules
`src/tests` folder may be confused with `tests/` where integration tests live. Shuttle tests check the specific functionality (protocol/infra) and they can live in the corresponding modules
