--- src/tests/mod.rs
@@ -1,2 +1,0 @@
-mod infra;
-mod protocol;
